+++ warp/native/warp.cpp
@@ -1013,6 +1013,8 @@ WP_API int wp_cuda_device_set_mempool_release_threshold(int ordinal, uint64_t th
 WP_API uint64_t wp_cuda_device_get_mempool_release_threshold(int ordinal) { return 0; }
 WP_API uint64_t wp_cuda_device_get_mempool_used_mem_current(int ordinal) { return 0; }
 WP_API uint64_t wp_cuda_device_get_mempool_used_mem_high(int ordinal) { return 0; }
+WP_API uint64_t wp_cuda_device_get_graph_mem_current(int ordinal) { return 0; }
+WP_API void wp_cuda_device_graph_mem_trim(int ordinal) { }
 WP_API void wp_cuda_device_get_memory_info(int ordinal, size_t* free_mem, size_t* total_mem) { }
 
 WP_API void* wp_cuda_context_get_current() { return NULL; }